Add files via upload
The project is related to create a Interface of ATM. We have all come across ATMs in our cities and it is built on Java. This complex project consists of five different classes and is a console-based application. When the system starts the user is prompted with user id and user pin. On entering the details successfully, then ATM functionalities are unlocked. The project allows to perform following operations:

1. Withdraw

2. Deposit

3. Transfer

4. Transactions History

5. Quit